Portfolio cleanup: security audit, README rewrite, JSDoc comments
- .env confirmed gitignored and never committed; .env.example updated
  with all five required variables and clear comments
- Removed debug console.log in resend.js that leaked the API key prefix
  into browser DevTools
- Added JSDoc file-level comments to BookingModal, StaffDashboard,
  Admin, and supabase.js explaining architecture decisions
- Full README rewrite: badges, architecture diagram, tech-stack table,
  What I Learned section, local setup guide, project tree
- Add api/send-email.js, src/lib/utils.js, supabase_rls_fix.sql, and
  test-e2e.mjs (all use env vars, no hardcoded credentials)
- Major feature additions: walk-in mobile-first flow, Brevo SMTP email,
  patient codes, staff analytics, no-show tracking, admin date blocking

+++ src/data/doctors.js
@@ -196,7 +196,7 @@ export function getAvailableDates(schedule) {
     const date = new Date(today);
     date.setDate(today.getDate() + i);
 
-    const dayName = date.toLocaleDateString('en-US', { weekday: 'long' });
+    const dayName = date.toLocaleDateString('en-US', { weekday: 'short' });
     const slot = schedule.find((s) => s.day === dayName);
 
     if (slot) {
